Cleanup warp related code
