+++ warp/handlers/signature_request_test.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"errors"
 	"testing"
-	"time"
 
 	"github.com/ava-labs/avalanchego/database/memdb"
 	"github.com/ava-labs/avalanchego/ids"
@@ -57,11 +56,9 @@ func TestMessageSignatureHandler(t *testing.T) {
 				require.EqualValues(t, 1, stats.messageSignatureRequest.Count())
 				require.EqualValues(t, 1, stats.messageSignatureHit.Count())
 				require.EqualValues(t, 0, stats.messageSignatureMiss.Count())
-				require.Greater(t, stats.messageSignatureRequestDuration.Value(), time.Duration(0))
 				require.EqualValues(t, 0, stats.blockSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.blockSignatureHit.Count())
 				require.EqualValues(t, 0, stats.blockSignatureMiss.Count())
-				require.EqualValues(t, stats.blockSignatureRequestDuration.Value(), time.Duration(0))
 			},
 		},
 		"unknown message": {
@@ -74,11 +71,9 @@ func TestMessageSignatureHandler(t *testing.T) {
 				require.EqualValues(t, 1, stats.messageSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.messageSignatureHit.Count())
 				require.EqualValues(t, 1, stats.messageSignatureMiss.Count())
-				require.Greater(t, stats.messageSignatureRequestDuration.Value(), time.Duration(0))
 				require.EqualValues(t, 0, stats.blockSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.blockSignatureHit.Count())
 				require.EqualValues(t, 0, stats.blockSignatureMiss.Count())
-				require.EqualValues(t, stats.blockSignatureRequestDuration.Value(), time.Duration(0))
 			},
 		},
 	}
@@ -159,11 +154,9 @@ func TestBlockSignatureHandler(t *testing.T) {
 				require.EqualValues(t, 0, stats.messageSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.messageSignatureHit.Count())
 				require.EqualValues(t, 0, stats.messageSignatureMiss.Count())
-				require.EqualValues(t, stats.messageSignatureRequestDuration.Value(), time.Duration(0))
 				require.EqualValues(t, 1, stats.blockSignatureRequest.Count())
 				require.EqualValues(t, 1, stats.blockSignatureHit.Count())
 				require.EqualValues(t, 0, stats.blockSignatureMiss.Count())
-				require.Greater(t, stats.blockSignatureRequestDuration.Value(), time.Duration(0))
 			},
 		},
 		"unknown block": {
@@ -176,11 +169,9 @@ func TestBlockSignatureHandler(t *testing.T) {
 				require.EqualValues(t, 0, stats.messageSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.messageSignatureHit.Count())
 				require.EqualValues(t, 0, stats.messageSignatureMiss.Count())
-				require.EqualValues(t, stats.messageSignatureRequestDuration.Value(), time.Duration(0))
 				require.EqualValues(t, 1, stats.blockSignatureRequest.Count())
 				require.EqualValues(t, 0, stats.blockSignatureHit.Count())
 				require.EqualValues(t, 1, stats.blockSignatureMiss.Count())
-				require.Greater(t, stats.blockSignatureRequestDuration.Value(), time.Duration(0))
 			},
 		},
 	}
